TIP-863: move Category Translation to Enrichment
