+++ UPGRADE.md
@@ -6,6 +6,10 @@ Several classes and services have been moved or renamed. The following commands 
 
 ```bash
 
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Entity\\CategoryTranslation/Akeneo\\Pim\\Enrichment\\Component\\Category\\Model\\CategoryTranslation/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Model\\CategoryTranslationInterface/Akeneo\\Pim\\Enrichment\\Component\\Category\\Model\\CategoryTranslationInterface/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Entity\\Category/Akeneo\\Pim\\Enrichment\\Component\\Category\\Model\\Category/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Model\\CategoryInterface/Akeneo\\Pim\\Enrichment\\Component\\Category\\Model\\CategoryInterface/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Model\\AbstractAssociation/Akeneo\\Pim\\Enrichment\\Component\\Product\\Model\\AbstractAssociation/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Model\\AssociationInterface/Akeneo\\Pim\\Enrichment\\Component\\Product\\Model\\AssociationInterface/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Model\\ProductModelAssociation/Akeneo\\Pim\\Enrichment\\Component\\Product\\Model\\ProductModelAssociation/g'
@@ -45,7 +49,7 @@ find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Model\
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\CatalogBundle\\Entity\\AttributeGroupTranslation/Akeneo\\Pim\\Structure\\Component\\Model\\AttributeGroupTranslation/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Model\\AttributeGroupTranslationInterface/Akeneo\\Pim\\Structure\\Component\\Model\\AttributeGroupTranslationInterface/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/pim_catalog.model.currency.interface/akeneo_channel.model.currency.interface/g'
-find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Validator\\Constraints\\ActivatedLocale/Akeneo\\Channel\\Component\\Validator\\Constraint\\ActivatedLocale`/g'
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Validator\\Constraints\\ActivatedLocale/Akeneo\\Channel\\Component\\Validator\\Constraint\\ActivatedLocale/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Validator\\Constraints\\Locale/Akeneo\\Channel\\Component\\Validator\\Constraint\\Locale/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\CatalogBundle\\Doctrine\\ORM\\Repository\\LocaleRepository/Akeneo\\Channel\\Bundle\\Doctrine\\Repository\\LocaleRepository/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Component\\Catalog\\Repository\\LocaleRepositoryInterface/Akeneo\\Channel\\Component\\Repository\\LocaleRepositoryInterface/g'
