+++ src/Akeneo/Pim/Enrichment/Component/Category/Model/CategoryTranslation.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Entity;
+namespace Akeneo\Pim\Enrichment\Component\Category\Model;
 
 use Akeneo\Tool\Component\Localization\Model\AbstractTranslation;
 use Akeneo\Pim\Enrichment\Component\Category\Model\CategoryTranslationInterface;
